feat: Created custom PhoSelect component (#142)
The standard select wasn't playing nice with the multiple flag. More
over, we could not style it freely.

The new component exposes an API matching the previous PhoSelect but it
extends it much more, by providing a dropdown-like select experience.

TODO: Handle keyboard inputs (so that users can navigate the PhoSelect
via the keyboard).